+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `0.1.1`
   * Improved documentation and added dev environment instructions
+  * Added function `updateTicket`
 
 * `0.1.0`
   * First working module with system health check